+++ RELEASENOTES.md
@@ -2,8 +2,9 @@
 
 ## Version 0.8.9
 Release date 2021-03-28
+- **Exits** new name of gables in module edit form.
 - **Bug fix** of issue #23 adding group owned stations.
-
+- **Bug fix** of rounding length and speed for modules.
 ## Version 0.8.8
 Release date 2021-03-24
 - **Cargo types** added with a first set of common items, translated to supported languages and coded according to NHM.